Move spec proxy behavior to optional modifiers.
HTTP proxies are required to implement certain functionality as per the
spec. However, some customers want to use Martian more as a transparent
proxy and would rather it not manipulate the request/response at all if
possible.

I have moved some of the core functionality such as checking for bad
framing and the removal of hop-by-hop headers to modifiers that will be
recommended, but not required.