Fix marbl test flake by adding a short sleep and failing fast.
The flake was caused by a race condition between subscribing the frame channel during websocket connection creation and the first write happening.  The test starts writing to the handler as soon as the websocket is connected, but there is a delay between the websocket being created and when a channel is created and subscribed.  If writes happen before the channel is subscribed, those writes don't get propagated to the websocket.  This should fail the test, but the test used t.Errorf(), which only records the failure and continues the test.  Because those initial writes were lost, the test ends up reading fewer messages than was sent and keeps waiting for more messages that will never arrive, thereby hanging the test.

The solution is to add a short sleep before the test begins writing to the handler.  Also changing t.Errorf() to t.Fatalf() makes the test fail fast and not wait for more messages.

This commit also fixes an unsafe map entry deletion in Write() that didn't hold a exclusive lock on the map.

Also added error logging in the rare instance that there's a map key ID collision.