If session is Hijacked close connection in handleLoop()
The Session.Hijack() doc says that "the connection will be closed following the return of the hijacker."
This is not exactly in the current code base.
If session is hijacked the handle() function returns nil, then handleLoop() tries to perform another loop, the loop fails on read t/o and only then the connection is closed.

This patch add a check if session is hijacked in handleLoop().
If the session is hijacked the connection is closed.