Add support for Server Sent Events
Server Sent Events [1] use chunked encoding to separate events i.e. each event is sent in a separate chunk.
To support proxying of Server Sent Events we must make sure that:
- we do preserve the original chunking,
- we flush after each chunk.

Given the net/http/internal/chunked.go implementations.
The first point is true if buffer size (32kB) is greater than the chunk (event) size in bytes.
This is because chunkedReader.Read() [2] returns a full chunk if it fits the passed read buffer and chunkedWriter.Write() returns io.ErrShortWrite if the data cannot be written in one Write() call.
The second point is addressed by flushing when chunk trailer ("\r\n") is written.

[1] https://developer.mozilla.org/en-US/docs/Web/API/Server-sent_events/Using_server-sent_events
[2] https://github.com/golang/go/blob/c85848a4a62f839427bb56cb2772e008e405cf1f/src/net/http/internal/chunked.go#L70
[3] https://github.com/golang/go/blob/c85848a4a62f839427bb56cb2772e008e405cf1f/src/net/http/internal/chunked.go#L203